- Added flow that handles adding and listing of reviews

+++ Domashna3/DiansWineryAppWIP/src/main/java/mk/ukim/finki/dianswineryappwip/model/Review.java
@@ -15,14 +15,14 @@ public class Review {
     private Winery winery;
 
     @Column(nullable = false)
-    private double rating;
+    private int rating;
 
     private String comment;
 
     public Review() {
     }
 
-    public Review(double rating, String comment) {
+    public Review(int rating, String comment) {
         this.rating = rating;
         this.comment = comment;
     }
